requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
python-dotenv
pandas
langfuse
requests
PyMuPDF
torch
